change structure of packages
